pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "bsb-nest"
authors = [{name = "Robin De Schepper", email = "robingilbert.deschepper@unipv.it"}]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
dynamic = ["version", "description"]
dependencies = ["bsb-core~=5.0"]

[tool.flit.module]
name = "bsb_nest"

[project.entry-points."bsb.simulation_backends"]
nest = "bsb_nest"

[project.optional-dependencies]
parallel = ["bsb-core[parallel]"]
test = [
    "bsb-core[parallel]",
    "bsb-test~=4.0",
    "bsb-hdf5~=5.0",
    # Required to load the Brunel config file
    "bsb-arbor~=4.0",
    "coverage~=7.0"
]
dev = [
    "bsb-nest[test]",
    "build~=1.0",
    "twine~=4.0",
    "pre-commit~=3.5",
    "black~=24.1.1",
    "isort~=5.12",
    "snakeviz~=2.1",
    "bump-my-version~=0.24"
]

[tool.black]
line-length = 90

[tool.isort]
profile = "black"
known_third_party = ["nest"]

[tool.bumpversion]
current_version = "4.3.2"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = "--no-verify"

[tool.bumpversion.parts.pre_l]
values = ["dev", "a", "b", "rc", "final"]
optional_value = "final"

[[tool.bumpversion.files]]
filename = "bsb_nest/__init__.py"